+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.7] - 2023-11-13
+
+### Added
+
+### Fixed
+
+- Add `isValidVat` and `supportedCountriesVat` functions.
+
+### Changed
+
+- Change `isValidEUVat` to `isValidViesVat`
+- Remove "vat" as an id doc type for `isValidIdDoc` function.
+
+### Removed
+
 ## [0.5.6] - 2023-11-13
 
 ### Added